(BOLT-643) Specify input method as stdin
Previously, we were using the default input method of "both", which
passes input over stdin and in environment variables. Because these
tasks can take very large parameters (especially upload and script),
they may fail to execute if the size of their params exceeds the
maximum allowable size for an environment variable. Because they don't
actually _use_ the environment variables, we now explicitly specify that
they should not be provided.