(BOLT-815) Mark task arguments that are file content as sensitive
The script and upload tasks take the file content to run or upload as a
string parameter. Previously, this content would be included in log
files and stored in the orchestrator database. This adds metadata for
the shim tasks to describe all the parameters, and marks the file
content parameters as sensitive to avoid storing or exposing them, as
well as avoid overwhelming the user with excessive output at debug
level.