(BOLT-830) Handle running scripts on windows targets
Previously running a script on windows targets would not work over the PCP transport. This commit adds logic in the `bolt_shim::script` task to handle running scripts on windows targets. This logic requires that the script file `name` (including extension) is sent along with the script `content` and `arguments`. In order to maintain backwards compatability this is optional metadata. An error will be raised if the task `name` metadata is not specified and the target OS is determined to be windows.