(BOLT-830) Only support powershell scripts for windows targets
Trying to find the correct interpreter for scripts other than powershell adds complexity to this task and is not likely to be very portable. This commit allows powershell scripts to be run and should satisfy user demand for now. A clean error message will be raised if the target system is windows and a non-powershell script is requested to be run.