(BOLT-1575) Make tmpdir in current directory, not /tmp
This updates the shim to create a temporary directory relative to the
directory that PXP agent uploads the tasks to, instead of in the system
tempdir. Using the system tempdir is an issue when users have the
directory mounted with `noexec`, which is required for CIS2 compliance
(and is a somewhat common configuration). In Bolt you can typically
configure a different tmpdir for SSH and WinRM transports to use, but
PCP doesn't have such a configuration option. Rather than making the
tmpdir configurable, this updates the shim to make a tmpdir relative to
`__dir__` which is the current directory.